+++ pygccxml/declarations/decl_printer.py
@@ -10,11 +10,12 @@ user friendly format
 
 import os
 import sys
-from . import calldef
+from . import calldef_members
 from . import algorithm
 from . import decl_visitor
 from . import variable_t
 from . import calldef_t
+from . import type_traits_classes
 from .. import utils
 
 
@@ -212,7 +213,7 @@ class decl_printer_t(decl_visitor.decl_visitor_t):
             "calling convention: __%s__" %
             decl.calling_convention +
             os.linesep)
-        if isinstance(decl, calldef.member_calldef_t):
+        if isinstance(decl, calldef_members.member_calldef_t):
             self.writer(indent +
                         "virtual: " +
                         str(decl.virtuality) +
@@ -243,7 +244,8 @@ class decl_printer_t(decl_visitor.decl_visitor_t):
         if self.__print_details:
             self.writer(indent +
                         'copy constructor: ' +
-                        str(self.__inst.is_copy_constructor) +
+                        str(type_traits_classes.is_copy_constructor(
+                            self.__inst)) +
                         os.linesep)
 
     def visit_destructor(self):